chore(agent-auth): CI green — split native_bridge tests under the line cap, ratchet ledger notes, ruff line length
